refactor(analysis): improve error handling and repo relevance scoring
- Introduce NoReposError class for better no repositories error signaling
- Replace error string checks with NoReposError instance checks in API route
- Move repository relevance scoring logic to dedicated repoRelevance module
- Use new repo relevance scoring function for sorting repositories
- Add try-catch around readme and releases fetch to handle failures gracefully
- Refactor developer analysis to use only enriched repositories for metrics
- Remove deprecated CommitWeek type from GitHub typings
- Extract GitHubError to centralized errors module for consistency
- Clean up unused imports and redundant code related to error handling
